Adelanto del punto 3
Se ha creado el metodo listar y agregar coche, ademas se agregó una clase menu para darle estetica y manejo al programa

+++ Concesionario/Coche.cs
@@ -16,7 +16,7 @@ namespace Concesionario
 
 
         //Construcctor
-        public Coche(int id, string marca, string modelo, int km, int precio)
+        public Coche(int id, string marca, string modelo, int km, double precio)
         {
             this._ID = id;
             this._Marca = marca;
